+++ cli/cli.js
@@ -52,6 +52,7 @@ if (verifiedConfigPath) {
 
 const rimraf = require("rimraf");
 const ora = require("ora");
+const glob = require("glob");
 
 const scrub = require("./lib/scrubber");
 const minify = require("./lib/minifier");
@@ -76,21 +77,28 @@ class DjinnJS {
             /** TODO: Add chalk */
             await this.preflightCheck();
             await this.createTempDirectory();
+            await this.createInjectionsDirectory();
             await this.resetOutputDirectories();
             await this.createOutputDirectories();
 
             if (!this.silent) {
                 spinner.text = "Scrubbing JavaScript imports";
             }
+            await this.dependencyInjection();
             await this.scrubScripts();
-            await this.injectConfigScriptVariables();
+            await this.injectConfig();
+            await this.injectRuntime();
 
             if (!this.silent) {
                 spinner.text = "Minifying JavaScript";
             }
             await this.minifyScript();
             await this.relocateServiceWorker();
-            await this.relocateBroadcastWorker();
+
+            if (!this.silent) {
+                spinner.text = "Relocating WebAssembly files";
+            }
+            await this.relocateWASM();
 
             if (!this.silent) {
                 spinner.text = "Relocating CSS files";
@@ -116,6 +124,25 @@ class DjinnJS {
             console.log(error);
             console.log("\n");
             process.exit(1);
+        }
+    }
+
+    async relocateWASM() {
+        const distDir = path.resolve(__dirname, "../dist");
+        const files = glob.sync(`${distDir}/**/*.wasm`);
+        const outDir = path.resolve(cwd, this.config.publicDir, this.config.outDir);
+        let relocated = 0;
+        for (let i = 0; i < files.length; i++) {
+            const filename = files[i].replace(/(.*[\/\\])/g, "");
+            fs.copyFile(files[i], `${outDir}/${filename}`, error => {
+                if (error) {
+                    throw error;
+                }
+                relocated++;
+                if (relocated === files.length) {
+                    return;
+                }
+            });
         }
     }
 
@@ -188,18 +215,6 @@ class DjinnJS {
         });
     }
 
-    relocateBroadcastWorker() {
-        return new Promise((resolve, reject) => {
-            const publicPath = path.resolve(cwd, this.config.publicDir);
-            fs.copyFile(path.resolve(__dirname, "../broadcaster-worker.min.js"), `${publicPath}/broadcaster-worker.min.js`, error => {
-                if (error) {
-                    reject(error);
-                }
-                resolve();
-            });
-        });
-    }
-
     relocateServiceWorker() {
         return new Promise((resolve, reject) => {
             const publicPath = path.resolve(cwd, this.config.publicDir);
@@ -220,6 +235,7 @@ class DjinnJS {
                     data = data.replace("REPLACE_WITH_NO_CACHE_PATTERN", this.config.noCachePattern);
                     data = data.replace("REPLACE_WITH_CACHEBUST_URL", this.config.cachebustURL);
                     data = data.replace("REPLACE_WITH_RESOURCE_PATTERN", this.config.resourcePattern);
+                    data = data.replace("REPLACE_WITH_PRECACHE_URL", this.config.precacheURL);
 
                     fs.writeFile(`${publicPath}/service-worker.js`, data, error => {
                         if (error) {
@@ -237,23 +253,57 @@ class DjinnJS {
         });
     }
 
-    injectConfigScriptVariables() {
+    injectConfig() {
         return new Promise((resolve, reject) => {
-            const runtimeFile = path.join(__dirname, "temp", "config.js");
-            fs.readFile(runtimeFile, (error, buffer) => {
+            const configFile = path.join(__dirname, "temp", "config.js");
+            fs.readFile(configFile, (error, buffer) => {
                 if (error) {
                     reject(error);
                 }
                 let data = buffer.toString();
                 data = data.replace("REPLACE_WITH_OUTPUT_DIR_NAME", this.config.outDir);
                 data = data.replace("REPLACE_WITH_GTAG_ID", this.config.gtagId);
-                data = data.replace('"REPLACE_WITH_PJAX_STATUS"', this.config.pjax);
-                data = data.replace('"REPLACE_WITH_PREFETCH_STATUS"', this.config.predictivePrefetching);
+                data = data.replace("REPLACE_WITH_PREFETCH_STATUS", this.config.prefetching);
                 data = data.replace('"REPLACE_WITH_FOLLOW_REDIRECT_STATUS"', this.config.followRedirects);
-                data = data.replace('"REPLACE_WITH_USE_PERCENTAGE"', this.config.usePercentage);
-                data = data.replace('"REPLACE_WITH_USE_SERVICE_WORKER"', `${this.config.serviceWorker ? true : false}`);
                 data = data.replace('"REPLACE_WITH_PAGE_JUMP_OFFSET"', this.config.pageJumpOffset);
                 data = data.replace("REPLACE_WITH_MINIMUM_CONNECTION", this.config.minimumConnection);
+                data = data.replace('"REPLACE_WITH_USE_SERVICE_WORKER"', this.config.serviceWorker ? true : false);
+                fs.writeFile(configFile, data, error => {
+                    if (error) {
+                        reject(error);
+                    }
+                    resolve();
+                });
+            });
+        });
+    }
+
+    injectRuntime() {
+        return new Promise((resolve, reject) => {
+            const runtimeFile = path.join(__dirname, "temp", "runtime.js");
+            fs.readFile(runtimeFile, (error, buffer) => {
+                if (error) {
+                    reject(error);
+                }
+                let data = buffer.toString();
+
+                if (this.config.pjax) {
+                    const pjaxScript = fs.readFileSync(path.join(__dirname, "static-injections", "pjax.js")).toString();
+                    data = data.replace('"REPLACE_WITH_PJAX_INJECTION";', pjaxScript);
+                } else {
+                    data = data.replace('"REPLACE_WITH_PJAX_INJECTION";', "");
+                }
+
+                if (this.config.serviceWorker) {
+                    const swScript = fs.readFileSync(path.join(__dirname, "static-injections", "service-worker.js")).toString();
+                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', swScript);
+                } else if (this.config.pjax) {
+                    const noServiceWorkerScript = fs.readFileSync(path.join(__dirname, "static-injections", "no-service-worker.js")).toString();
+                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', noServiceWorkerScript);
+                } else {
+                    data = data.replace('"REPLACE_WITH_SERVICE_WORKER_INJECTION";', "");
+                }
+
                 fs.writeFile(runtimeFile, data, error => {
                     if (error) {
                         reject(error);
@@ -262,6 +312,45 @@ class DjinnJS {
                 });
             });
         });
+    }
+
+    injectDjinnjax() {
+        return new Promise((resolve, reject) => {
+            const djinnjax = path.resolve(cwd, "./node_modules/djinnjax/dist/djinnjax.js");
+            if (!fs.existsSync(djinnjax)) {
+                reject("The djinnjax package is not installed. Disable pjax or run 'npm i -S djinnjax'");
+            } else {
+                glob(`${path.resolve(djinnjax, "../")}/*.js`, (error, files) => {
+                    if (error) {
+                        reject(error);
+                    }
+                    const tempDir = path.join(__dirname, "injections");
+                    let count = 0;
+                    for (let i = 0; i < files.length; i++) {
+                        const filename = files[i].replace(/(.*[\/\\])/g, "");
+                        fs.copyFile(files[i], `${tempDir}/${filename}`, error => {
+                            if (error) {
+                                reject(error);
+                            }
+                            count++;
+                            if (count === files.length) {
+                                resolve();
+                            }
+                        });
+                    }
+                });
+            }
+        });
+    }
+
+    async dependencyInjection() {
+        try {
+            if (this.config.pjax) {
+                await this.injectDjinnjax();
+            }
+        } catch (error) {
+            throw error;
+        }
     }
 
     scrubScripts() {
@@ -310,8 +399,26 @@ class DjinnJS {
 
     cleanup() {
         return new Promise(resolve => {
-            rimraf.sync(path.join(__dirname, "temp"));
+            const tempPath = path.join(__dirname, "temp");
+            const injectionsPath = path.join(__dirname, "injections");
+            if (fs.existsSync(tempPath)) {
+                rimraf.sync(tempPath);
+            }
+            if (fs.existsSync(injectionsPath)) {
+                rimraf.sync(injectionsPath);
+            }
             resolve();
+        });
+    }
+
+    createInjectionsDirectory() {
+        return new Promise((resolve, reject) => {
+            fs.mkdir(path.join(__dirname, "injections"), error => {
+                if (error) {
+                    reject(error);
+                }
+                resolve();
+            });
         });
     }
 
@@ -326,13 +433,8 @@ class DjinnJS {
         });
     }
 
-    preflightCheck() {
-        return new Promise(resolve => {
-            if (fs.existsSync(path.join(__dirname, "temp"))) {
-                rimraf.sync(path.join(__dirname, "temp"));
-            }
-            resolve();
-        });
+    async preflightCheck() {
+        await this.cleanup();
     }
 }
 new DjinnJS(customConfig);