+++ cli/lib/validate-config.js
@@ -130,25 +130,24 @@ module.exports = function valdiateConfig(config, customConfig) {
             process.exit(1);
     }
 
-    switch (typeof customConfig.predictivePrefetching) {
-        case "boolean":
-            config.predictivePrefetching = customConfig.predictivePrefetching;
+    switch (typeof customConfig.prefetching) {
+        case "string":
+            switch (customConfig.prefetching) {
+                case "opt-out":
+                    config.prefetching = customConfig.prefetching;
+                    break;
+                case "opt-in":
+                    config.prefetching = customConfig.prefetching;
+                    break;
+                default:
+                    console.log(`Invalid DjinnJS configuration. The prefetching value must be "opt-in" or "opt-out".`);
+                    process.exit(1);
+            }
             break;
         case "undefined":
             break;
         default:
             console.log(`Invalid DjinnJS configuration. The predictivePrefetching value must be a boolean.`);
-            process.exit(1);
-    }
-
-    switch (typeof customConfig.usePercentage) {
-        case "boolean":
-            config.usePercentage = customConfig.usePercentage;
-            break;
-        case "undefined":
-            break;
-        default:
-            console.log(`Invalid DjinnJS configuration. The usePercentage value must be a boolean.`);
             process.exit(1);
     }
 
@@ -199,6 +198,17 @@ module.exports = function valdiateConfig(config, customConfig) {
             break;
         default:
             console.log(`Invalid DjinnJS configuration. The minimumConnection value must be a string.`);
+            process.exit(1);
+    }
+
+    switch (typeof customConfig.precacheURL) {
+        case "string":
+            config.precacheURL = customConfig.precacheURL;
+            break;
+        case "undefined":
+            break;
+        default:
+            console.log(`Invalid DjinnJS configuration. The precacheURL value must be a string.`);
             process.exit(1);
     }