+++ cli/static-injections/no-service-worker.js
@@ -1,0 +1,2 @@
+const event = new CustomEvent("pjax:init");
+document.dispatchEvent(event);