+++ src/core/config.ts
@@ -1,9 +1,7 @@
 export const djinnjsOutDir = "REPLACE_WITH_OUTPUT_DIR_NAME";
 export const gaId = "REPLACE_WITH_GTAG_ID";
-export const usePjax = "REPLACE_WITH_PJAX_STATUS";
-export const doPrefetching = "REPLACE_WITH_PREFETCH_STATUS";
+export const prefetching = "REPLACE_WITH_PREFETCH_STATUS";
 export const followRedirects = "REPLACE_WITH_FOLLOW_REDIRECT_STATUS";
-export const usePercentage = "REPLACE_WITH_USE_PERCENTAGE";
-export const useServiceWorker = "REPLACE_WITH_USE_SERVICE_WORKER";
 export const pageJumpOffset = "REPLACE_WITH_PAGE_JUMP_OFFSET";
 export const minimumConnection = "REPLACE_WITH_MINIMUM_CONNECTION";
+export const useServiceWorker = "REPLACE_WITH_USE_SERVICE_WORKER";