+++ src/core/types.d.ts
@@ -1,13 +1,1 @@
-export type PjaxResources = {
-    eager: Array<string>;
-    lazy: Array<string>;
-};
-
-export type WorkerResponse = {
-    type: "eager" | "lazy" | "parse";
-    files: Array<string>;
-    requestUid: string | null;
-    pjaxFiles: PjaxResources;
-};
-
 export type WebComponentLoad = null | "lazy" | "eager";