+++ src/service-workers/offline-first.js
@@ -181,6 +181,37 @@ async function bustContent(url) {
 }
 
 async function cachebust(url) {
-    bustResources();
-    bustContent(url);
+    await bustResources();
+    await bustContent(url);
+    await warmTheCache();
+}
+
+async function warmTheCache() {
+    const precacheURL = "REPLACE_WITH_PRECACHE_URL";
+    if (precacheURL.length) {
+        const request = await fetch(precacheURL, {
+            cache: "no-cache",
+            credentials: "include",
+            headers: new Headers({
+                Accept: "application/json",
+            }),
+        });
+        const response = await request.json();
+        if (request.ok) {
+            if (response?.resources) {
+                event.waitUntil(
+                    caches.open(resourcesCacheId).then(cache => {
+                        return cache.addAll(response.resources);
+                    })
+                );
+            }
+            if (response?.content) {
+                event.waitUntil(
+                    caches.open(contentCacheId).then(cache => {
+                        return cache.addAll(response.content);
+                    })
+                );
+            }
+        }
+    }
 }